requirements.txt
phidata
packaging
openai
streamlit
flask